+++ ext/polyphony/backend_libev.c
@@ -759,9 +759,9 @@ void Backend_timeout_callback(EV_P_ ev_timer *w, int revents)
 
 VALUE Backend_timeout(int argc,VALUE *argv, VALUE self) {
   VALUE duration;
-  VALUE exception_class;
+  VALUE exception;
   VALUE move_on_value = Qnil;
-  rb_scan_args(argc, argv, "21", &duration, &exception_class, &move_on_value);
+  rb_scan_args(argc, argv, "21", &duration, &exception, &move_on_value);
 
   Backend_t *backend;
   struct libev_timeout watcher;
@@ -778,9 +778,8 @@ VALUE Backend_timeout(int argc,VALUE *argv, VALUE self) {
   result = rb_ensure(Backend_timeout_ensure_safe, Qnil, Backend_timeout_ensure, (VALUE)&timeout_ctx);
   
   if (result == timeout) {
-    if (exception_class == Qnil) return move_on_value;
-    VALUE exception = rb_funcall(exception_class, ID_new, 0);
-    RAISE_EXCEPTION(exception);
+    if (exception == Qnil) return move_on_value;
+    RAISE_EXCEPTION(backend_timeout_exception(exception));
   }
 
   RAISE_IF_EXCEPTION(result);